refactor(im_adapter): 创建 im_adapter 核心模块，统一渲染层
Source: issue #886
Type: refactor
