+++ src/im_adapter/streaming.rs
